requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
click==8.3.1
eval_type_backport==0.3.0
fastapi==0.121.2
functions==0.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
invoke==2.2.1
mistralai==1.9.11
numpy==2.3.5
packaging==25.0
pluggy==1.6.0
psycopg2==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
schemas==0.7.1
six==1.17.0
sniffio==1.3.1
spglib==2.6.0
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
utils==1.0.2
uvicorn==0.38.0
